classes ディレクトリ下のファイルを分別、defaultWWAData を classes 下に移動
